chore: make logout optional when changing password (#11335)
https://linear.app/unleash/issue/SA-87/make-logout-optional-when-changing-password

Make logout optional when changing password.

This allows us to use this method when setting a new password in the
signup dialog without forcing a logout, improving UX.

+++ src/lib/services/user-service.ts
@@ -573,7 +573,11 @@ export class UserService {
         return user;
     }
 
-    async changePassword(userId: number, password: string): Promise<void> {
+    async changePassword(
+        userId: number,
+        password: string,
+        { logoutUser }: { logoutUser?: boolean } = { logoutUser: true },
+    ): Promise<void> {
         this.validatePassword(password);
         const passwordHash = await bcrypt.hash(password, saltRounds);
 
@@ -582,8 +586,11 @@ export class UserService {
             passwordHash,
             disallowNPreviousPasswords,
         );
-        await this.sessionService.deleteSessionsForUser(userId);
         await this.resetTokenService.expireExistingTokensForUser(userId);
+
+        if (logoutUser) {
+            await this.sessionService.deleteSessionsForUser(userId);
+        }
     }
 
     async changePasswordWithPreviouslyUsedPasswordCheck(
